Move `euc_bit` function to `metric.py`
This function will be part of `metric.py` and having it in
testing module is not a good practice.
